fix: align merge conflicts
